requirements.txt
rq==1.0
redis==3.2.1
requests==2.21.0
dnspython==1.16.0
geoip2==2.9.0
pyaml==19.4.1
pycryptodome==3.9.0
ecdsa==0.13.2